Revived the intended pattern of converting GraphicsContext to ComputeContext to do compute work.
The full intent of the APIs is that they also provide methods for setting root constants and descriptor tables and such that automatically call the appropriate methods on the command list given the context.
That part still isn't ported over, but at least functions don't need to care about what type of context the caller has anymore.
(ParticleSystem::Update for instance now automatically adapts to running on async compute if that's what you provide it.)